Move DataModel para diretório de abstração
