+++ samples/UsingData/BlogDataModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Linq.Expressions;
-using E5R.Architecture.Data;
+using E5R.Architecture.Data.Abstractions;
 
 namespace UsingData
 {
