+++ src/E5R.Architecture.Data/DataReducer`1.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
+using E5R.Architecture.Data.Abstractions;
 
 namespace E5R.Architecture.Data
 {
